# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Approve": {
        "additionalProperties": false,
        "description": "Approve the checkpoint as it is, with a whole edited output, or with edits.",
        "properties": {
          "competitors": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitorEdits"
              },
              {
                "type": "null"
              }
            ]
          },
          "edited_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Output"
          },
          "pain_points": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PainPointEdits"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "Approve",
        "type": "object"
      },
...
      "ChangelogAlerts": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/ChangelogAlert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "items_matched": {
            "title": "Items Matched",
            "type": "integer"
          },
          "run_id": {
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "alerts",
          "items_matched"
        ],
        "title": "ChangelogAlerts",
        "type": "object"
      },
...
      "CreateRun": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "$ref": "#/components/schemas/RunInput"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "input"
        ],
        "title": "CreateRun",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PricingAlert": {
        "additionalProperties": false,
        "description": "A pricing page that changed between two snapshots, and how.",
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/PricingChange"
            },
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "previous_fetched_at": {
            "format": "date-time",
            "title": "Previous Fetched At",
            "type": "string"
          },
          "product_id": {
            "pattern": "^prod_[A-Za-z0-9]+$",
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "minLength": 1,
            "title": "Product Name",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "url": {
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "product_name",
          "url",
          "previous_fetched_at",
          "detected_at",
          "changes"
        ],
        "title": "PricingAlert",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Resume": {
        "additionalProperties": false,
        "properties": {
          "from_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Stage"
          }
        },
        "title": "Resume",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "pending",
          "running",
          "awaiting_approval",
          "paused_quota",
          "failed",
          "completed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "RunSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idea": {
            "title": "Idea",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "idea",
          "created_at",
          "updated_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
      "RunView": {
        "properties": {
          "awaiting_approval": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awaiting Approval"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idea": {
            "title": "Idea",
            "type": "string"
          },
          "input": {
            "$ref": "#/components/schemas/RunInput"
          },
          "pause_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Reason"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageView"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "idea",
          "created_at",
          "updated_at",
          "input",
          "seed",
          "error",
          "pause_reason",
          "awaiting_approval",
          "stages"
        ],
        "title": "RunView",
        "type": "object"
      },
...
      "StageOutputView": {
        "properties": {
          "edited": {
            "title": "Edited",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "original_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Output"
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          }
        },
        "required": [
          "key",
          "status",
          "schema_version",
          "edited",
          "output",
          "original_output"
        ],
        "title": "StageOutputView",
        "type": "object"
      },
      "StageStatus": {
        "enum": [
          "pending",
          "running",
          "awaiting_approval",
          "failed",
          "completed"
        ],
        "title": "StageStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Start, watch and approve runs. Long stages run in the queue worker.",
    "title": "ProductFoundry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pricing/alerts": {
      "get": {
        "description": "Changes found between two snapshots of a tracked pricing page, newest first.",
        "operationId": "list_pricing_alerts_pricing_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PricingAlert"
                  },
                  "title": "Response List Pricing Alerts Pricing Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pricing Alerts"
      }
    },
    "/runs": {
      "get": {
        "description": "Every run, newest first.",
        "operationId": "list_runs_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs"
      },
      "post": {
        "description": "Create a run and queue it. It stops at the first checkpoint.",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRun"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Run"
      }
    },
    "/runs/{run_id}": {
      "get": {
        "description": "A run with the status of each stage. `paused_quota` is a pause, not a failure.",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Run"
      }
    },
    "/runs/{run_id}/approve": {
      "post": {
        "description": "Approve the open checkpoint, unchanged or edited, and queue the rest of the run.",
        "operationId": "approve_runs__run_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Approve"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Approve"
      }
    },
    "/runs/{run_id}/changelog": {
      "get": {
        "description": "What competitors shipped that touches this run: fixes for its pain points, and\nnew features its PRD does not cover. Matching is done by `changelog match`.",
        "operationId": "get_changelog_alerts_runs__run_id__changelog_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangelogAlerts"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Changelog Alerts"
      }
    },
    "/runs/{run_id}/exports/{name}": {
      "get": {
        "description": "The pain-point report, the PRD or the task plan, as Markdown or JSON.",
        "operationId": "download_export_runs__run_id__exports__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "enum": [
                "report",
                "prd",
                "tasks"
              ],
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "md",
              "enum": [
                "md",
                "json"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/markdown": {}
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Download Export"
      }
    },
    "/runs/{run_id}/resume": {
      "post": {
        "description": "Queue a paused or failed run again, or run it again from a named stage.",
        "operationId": "resume_runs__run_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Resume"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Resume"
      }
    },
    "/runs/{run_id}/stages/{stage_key}": {
      "get": {
        "operationId": "get_stage_output_runs__run_id__stages__stage_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stage_key",
            "required": true,
            "schema": {
              "title": "Stage Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageOutputView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Stage Output"
      }
    }
  }
}